add handling of service unit group mappings to copy groups
In order to make service units stick, it needs to be handled via groups
